lots of changes

+++ install.sh
@@ -1,20 +1,25 @@
 #!/bin/bash
 
 source "$HOME/.dotfiles/install/declarations"
-source "$INSTALL/essential"
-source "$INSTALL/app-install"
-source "$INSTALL/programming-tools"
-source "$INSTALL/system-setup"
-source "$INSTALL/startup-scripts"
 
-if [ "$(uname)" == "Darwin" ]; then
+if [ "$(uname)" == "Linux" ]; then
+  if (( ${quick:?} )); then
+    source "$INSTALL/system-setup"
+  else
+    source "$INSTALL/essential"
+    source "$INSTALL/app-install"
+    source "$INSTALL/programming-tools"
+    source "$INSTALL/system-setup"
+    source "$INSTALL/startup-scripts"
+  fi
+  sudo apt-get -y --force-yes autoremove
+  sudo apt-get -y --force-yes update
+  sudo apt-get -y --force-yes upgrade
+elif [ "$(uname)" == "Darwin" ]; then
     source "$INSTALL/brew-setup"
     source "$INSTALL/osx-setup"
 fi
 
-sudo apt-get -y --force-yes autoremove
-sudo apt-get -y --force-yes update
-sudo apt-get -y --force-yes upgrade
 [ -d ~/tmp ]  &&  rm -rf ~/tmp
 
 clear
